fix: 이미지 변경 기능 취소시 이전 이미지로 변경 가능하도록 로직 수정 및 관련 파일 경로 수정
