+++ src/services/mypageUpdateService.ts
@@ -1,12 +1,5 @@
 import { db } from "@/firebase";
-import { uidType } from "@/types";
 import { doc, updateDoc } from "firebase/firestore";
-
-export const updateImageURL = async (uid: uidType, imgURL: string) => {
-	const userID = String(uid);
-	const docRef = doc(db, "user", `${userID}`);
-	await updateDoc(docRef, { photoURL: imgURL });
-};
 
 export const updateUserInfoData = async (
 	uid: string,
@@ -15,6 +8,7 @@ export const updateUserInfoData = async (
 		name: string;
 		team: string;
 		grade: string;
+		photoURL: string;
 	},
 ) => {
 	const docRef = doc(db, "user", uid);
@@ -23,5 +17,6 @@ export const updateUserInfoData = async (
 		name: userInfoData.name,
 		team: userInfoData.team,
 		grade: userInfoData.grade,
+		photoURL: userInfoData.photoURL,
 	});
 };
